Issue #4437: Modified RedundantImportCheckTest.java and moved its input files to the redundantimport subdirectory
