+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/imports/redundantimport/InputRedundantImport_UnnamedPackage.java
