+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/redundantimport/InputImportBug.java
@@ -1,0 +1,12 @@
+package com.puppycrawl.tools.checkstyle.checks.imports.redundantimport;
+
+import java.net.URL;
+
+public class InputImportBug {
+   //same as a class name
+   private static String URL = "This is a String object";
+
+   public InputImportBug() throws Exception {
+       URL url = new URL("file://this.is.a.url.object");
+   }
+}
